tools/release-scripts/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
feedgen==1.0.0
    # via -r requirements.in
idna==3.10
    # via requests
lxml==6.0.0
    # via
    #   -r requirements.in
    #   feedgen
markdown-it-py==3.0.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
python-dateutil==2.9.0.post0
    # via feedgen
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.4
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
urllib3==2.5.0
    # via requests
